pyproject.toml
[project]
name = "library"
version = "0.1.0"
description = "This is a Python package for the SWE API."
readme = "README.md"
authors = [
    { name = "Ole Menke", email = "meol1013@h-ka.de" },
    { name = "Paul Braeuninger", email = "brpa1033@h-ka.de" },
]
maintainers= [
    { name = "Ole Menke", email = "meol1013@h-ka.de" },
    { name = "Paul Braeuninger", email = "brpa1033@h-ka.de" },
]
requires-python = ">=3.14"
dependencies = [
    "anyio>=4.12.1",
    "cffi>=2.0.0",
    "cryptography>=46.0.7",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.135.1",
    "httpx[http2]>=0.28.1",
    "jwcrypto>=1.5.7",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "orjson>=3.11.7",
    "prometheus-client>=0.24.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg-binary>=3.3.3",
    "psycopg-pool>=3.3.0",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic-core>=2.41.5",
    "pydantic[email]>=2.12.5",
    "pyfiglet>=1.0.4",
    "python-dotenv>=1.2.2",
    "python-keycloak>=7.1.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "starlette>=0.52.1",
    "strawberry-graphql>=0.312.3",
    "tabulate>=0.10.0",
    "urllib3>=2.6.3",
    "uvicorn>=0.42.0",
]

[project.scripts]
library = "library:main"

[build-system]
requires = ["uv_build==0.11.6"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "fastapi[standard]>=0.135.1",
    "gevent>=25.9.1",
    "httptools>=0.7.1",
    "itsdangerous>=2.2.0",
    "locust>=2.43.3",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-env>=1.6.0",
    "pytest-html>=4.2.0",
    "pytest-mock>=3.15.1",
    "werkzeug>=3.1.7",
]
doc = [
    "griffe>=2.0.1",
    "markdown>=3.10.2",
    "mistune>=3.2.0",
    "mkdocs>=1.6.1",
    "mkdocs-build-plantuml-plugin>=2.1.0",
    "mkdocs-material>=9.7.6",
    "mkdocstrings[python]>=1.0.3",
    "mkdocstrings-python>=2.0.3",
]
lint = [
    "pip-audit>=2.10.0",
    "pysentry-rs>=0.4.3",
    "ruff>=0.15.7",
    "ty>=0.0.24",
    "types-openpyxl>=3.1.5.20260322",
    "types-tabulate>=0.10.0.20260308",
]

[tool.pytest.ini_options]
minversion = "8.4.1"
pythonpath = ["src"]
testpaths = ["tests/integration"]
addopts = "--html=report.html --cov --cov-report=html"
markers = [
    "rest: REST-Schnittstelle",
    "get_request: GET-Requests",
    "post_request: POST-Requests",
    "put_request: PUT-Requests",
    "delete_request: DELETE-Requests",
    "login: Login mit der REST-Schnittstelle",
    "health: Health des Servers ueberpruefen",
    "graphql: GraphQL-Schnittstelle",
    "query: Queries bei GraphQL",
    "mutation: Mutations bei GraphQL",
    "simple: Einfache Tests mit True und False",
    "unit",
    "unit_find_by_id",
    "unit_find",
    "unit_create",
]

[tool.pytest_env]
# LOGURU_LEVEL = "INFO"

[tool.ruff]
line-length = 132

[tool.ruff.lint]
select = [
    "FAST",
    "B",
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "A",
    "COM",
    "C4",
    "ISC001",
    "PIE",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "TID",
    "TC",
    "ARG",
    "PTH",
    "I",
    "N",
    "PERF",
    "E",
    "W",
    "D",
    "F",
    "PL",
    "UP",
    "FURB",
    "RUF",
]
ignore = [
    "D203",
    "D213",
    "RUF012",
    "COM812",
    "PLR6301",
    "TC001",
    "TC002",
    "TC003",
]

[tool.ruff.lint.isort]
case-sensitive = true

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.coverage.run]
include = ["src/library/service/*"]

[tool.ty.environment]
python = "./.venv"
root = ["./src", "./tests/integration"]
# FIXME: Add "./tests/unit" back, when the unit tests are written.

[tool.uv]
default-groups = "all"
